Add files via upload
Exercise 12.3 in the Game Theory by Steven Tadelis. 
Armed Conﬂict: Consider the following strategic situation: Two rival armies plan to seize a disputed territory. Each army’s general can choose either to attack (A) or to not attack (N). In addition, each army is either strong (S) or weak (W) with equal probability, and the realizations for each army are independent. Furthermore the type of each army is known only to that army’s general. An army can capture the territory if either (i) it attacks and its rival does not or (ii) it and its rival attack, but it is strong and the rival is weak. If both attack and are of equal strength then neither captures the territory. As for payoffs, the territory is worth m if captured and each army has a cost of ﬁghting equal to s if it is strong and w if it is weak, where s<w. If an army attacks but its rival does not, no costs are borne by either side. Identify all the pure-strategy Bayesian Nash equilibria of this game for the following two cases, and brieﬂy describe the intuition for your results: 

The action spaces of the two armies are A1 = A2 ={A, N}. (A  = attack, N = not attack)
The type spaces of the two armies are T1 = T2 = {S, W}. (S = strong, W = weak). Since, each army is either strong (S) or weak (W) with equal probability, Pr(S) = Pr(W) = ½. 
The strategy sets are S1 = S2 = {AA, AN, NA, NN} (AA = always attack, AN = attack if strong, but not attack if weak, NA = attack if weak, but not if strong, NN = never attack)


This is a brute force method to solve for the payoff matrix. 
To calculate the payoffs without computing the formulas use if statements to check who would win, in each scenario.